+++ weblate/static/editor/full.js
@@ -495,7 +495,7 @@
     }
 
     renderTranslation(el, service) {
-      var row = $("<tr/>").attr("class", "js-copy-machinery").data("raw", el);
+      var row = $("<tr/>").data("raw", el);
       row.append(
         $("<td/>")
           .attr("class", "target machinery-text")