+++ docs/changes.rst
@@ -27,6 +27,7 @@ Not yet released.
 * Docker container automatically enables TLS/SSL for outgoing e-mail when needed.
 * Added support for searching for resolved comments.
 * Added support for borgbackup 1.2.
+* Fixed applying of :guilabel:`Automatically translated` label.
 
 Weblate 4.10.1
 --------------